% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{add_node}
\alias{add_node}
\title{add_node}
\usage{
add_node(x1, y1, height = 5, width = 5, geom = "rect")
}
\arguments{
\item{x1}{x1 value}

\item{y1}{y1 value}

\item{height}{height}

\item{width}{width}

\item{geom}{geometric shape}
}
\description{
add node
}
