% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{add_rect}
\alias{add_rect}
\title{add_rect}
\usage{
add_rect(x1, y1, height = 5, width = 5)
}
\arguments{
\item{x1}{x1}

\item{y1}{y1}

\item{height}{height}

\item{width}{width}
}
\description{
add node in rectangle shape
}
