% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{addlabel}
\alias{addlabel}
\title{addlabel}
\usage{
addlabel(
  x1 = 0,
  y1 = 0,
  x2 = 10,
  y2 = 10,
  height = 3,
  width = 5,
  start = "ellipse",
  end = "rect",
  label = label,
  useLabel,
  ...
)
}
\arguments{
\item{x1}{x1}

\item{y1}{y1}

\item{x2}{x2}

\item{y2}{y2}

\item{height}{height}

\item{width}{width}

\item{start}{start}

\item{end}{end}

\item{label}{label}

\item{useLabel}{a boolean value}
}
\description{
add label to the ggobject
}
