% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{addline}
\alias{addline}
\title{addline}
\usage{
addline(
  x1 = 0,
  y1 = 0,
  x2 = 10,
  y2 = 10,
  height = 3,
  width = 5,
  start = "ellipse",
  end = "rect",
  linetype = "solid",
  size = 0.2,
  colour = "black",
  ...
)
}
\arguments{
\item{x1}{default x1 is 0}

\item{y1}{default y1 is 0}

\item{x2}{default x2 is 10}

\item{y2}{default y2 is 10}

\item{height}{default height is 3}

\item{width}{default width is 5}

\item{start}{the default start shape is ellipse}

\item{end}{the default end shape is rectangle}

\item{linetype}{the type of the line to use}

\item{size}{the width of the line, default 0.2}

\item{colour}{the color of the line, default black}
}
\value{
a ggplot object of the line
}
\description{
Add line to the mediation plot
}
