% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{addLinetype}
\alias{addLinetype}
\title{addLinetype}
\usage{
addLinetype(res, pval = 0.05)
}
\arguments{
\item{res}{the information of the mediation relationship}

\item{pval}{the pvalue used to judge the significance. pvalue < pval, significant, pvalue > pval, unsignificant.}
}
\value{
the information with annotated line type to use
}
\description{
add line

Judging line types to use according to the result. If relationship is unsignificant, use dashed line, otherwise, use solid line.
}
