% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{addpos}
\alias{addpos}
\title{addpos}
\usage{
addpos(df)
}
\arguments{
\item{df}{the result used for plot the sem mediation relationship}
}
\value{
the data.frame with position information added
}
\description{
add position
}
