% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{adjustPos}
\alias{adjustPos}
\title{adjustPos}
\usage{
adjustPos(df, maxx = 80, maxy = 30, height = 3, width = 5)
}
\arguments{
\item{df}{the data.frame of the information extracted from fitted model}

\item{maxx}{the max value of x, default 80}

\item{maxy}{the max value of y, default 30}

\item{height}{the height of the figure, default 3}

\item{width}{the width of the figure, default 5}
}
\value{
the data.frame with adjusted positions
}
\description{
adjust the position of the data.frame
}
