% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_scObj.R
\name{bootstrapSO}
\alias{bootstrapSO}
\title{bootstrapSO}
\usage{
bootstrapSO(
  iso,
  feature.name = "seurat_clusters",
  minCNUM.tot = 10,
  sampleProp = 0.1,
  minCNUM.new = 10,
  runs = 5,
  suggestedRunTimes = F,
  returnNameOnly = T
)
}
\arguments{
\item{iso}{A seurat object}

\item{feature.name}{The name of the feature, which come from the metatable of seurat object, to classfiy the cells}

\item{minCNUM.tot}{The minimum cell number for a cluster, cluster with cells less the minCNUM.tot will be removed from the analysis. Default 10.}

\item{sampleProp}{The porportion of cells used in resampling for each subcluster. Default 0.1.}

\item{minCNUM.new}{Default is 10. The expected minimum cells for resmple. If estimated cells used for resample, which is the cluster_cell_size*sampleProp, is smaller than the minimum cell number for resampled cells in each subcluster, use minimum cell number as the resample size. This value should be not greater than minCNUM.tot to avoid error.}

\item{runs}{Total runs for the bootstrap. Default is 5.}

\item{suggestedRunTimes}{default is False. If is True, the bootstrap won't run. The function will estimate the maximum runs for non-redundancy sampling results.}

\item{returnNameOnly}{default is True. If is True, only return the cellnames rather than the resampled objects.}
}
\value{
a seurat object list or the cell names of a seurat object list or the estimated maximum run times.
}
\description{
bootstrap seurat object
}
