% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{clear_scCellMediator_workspace}
\alias{clear_scCellMediator_workspace}
\title{clear_scCellMediator_workspace()}
\usage{
clear_scCellMediator_workspace()
}
\description{
identify and clear the work space of temporal files generated
}
