% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Convex Hull}
\usage{
convex_hull(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates, a n by 2 matrix, or anything that
can be coerced to a matrix via \code{as.matrix}.}

\item{y}{a numeric vector of y coordinates, only needed if x is a numeric
vector}
}
\value{
a list
}
\description{
return the convex hull for a set of (x, y) points
}
\examples{
## Example 1
# Default, x and y numeric vectors
x <- rnorm(150)
y <- rnorm(150)

hull <- convex_hull(x, y)

# plot the data and the hull
plot(x, y) 
lines(hull)

# show interior points in red
points(x[-attr(hull, "indices")], 
       y[-attr(hull, "indices")], 
       col = "red", pch = 16) 

## Example 2
# With a matrix
mat <- matrix(rnorm(500), nrow = 250)
hull2 <- convex_hull(mat)

plot(mat)
lines(hull2)

# show interior points in red
points(mat[-attr(hull2, "indices"), ], col = "red", pch = 16)

## Example 3
# Two columns from the mtcars data set
hull3 <- convex_hull(mtcars[, c("wt", "mpg")])

plot(mtcars[, c("wt", "mpg")])
lines(hull3)
points(mtcars[-attr(hull3, "indices"), c("wt", "mpg")], col = "red", pch = 16)

}

