% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{countM}
\alias{countM}
\title{countM}
\usage{
countM(group)
}
\arguments{
\item{group}{a vector contains all names of the components in the mediation model}
}
\value{
the number of M (mediators)
}
\description{
count the number of M (mediators)
}
