% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{createDir}
\alias{createDir}
\title{createDir}
\usage{
createDir(dirpath)
}
\arguments{
\item{dirpath}{The directory path to check.}
}
\description{
If directory not exist, create the directory. Otherwise, ignore the command.
}
