% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{detect_suspecious_clusters}
\alias{detect_suspecious_clusters}
\title{detect_suspecious_clusters}
\usage{
detect_suspecious_clusters(final_signal_table)
}
\arguments{
\item{final_signal_table}{a table per row shows: ct, gene, cluster, pos (gene expression in target cluster), neg (gene expression in control cells), cnum (cell number detected in the cluster)}
}
\value{
a vector of elements shows cell type and cluster glued by comma
}
\description{
Detecting suspecious clusters
}
