% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{distributePos}
\alias{distributePos}
\title{distributePos}
\usage{
distributePos(start, end, count)
}
\arguments{
\item{start}{the start position}

\item{end}{the end position}

\item{count}{the number of the points to segment the postion}
}
\value{
a number vector show the accurate positions
}
\description{
distributed the position
}
