% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{estimate_pretreated_so_by_refMarkers}
\alias{estimate_pretreated_so_by_refMarkers}
\title{estimate_pretreated_so_by_refMarkers estimate pretreated seurat object by given reference markers}
\usage{
estimate_pretreated_so_by_refMarkers(
  so,
  refList,
  clusterMarkerPath,
  tissue = NULL,
  annotateName = "CellType",
  clusterName = "seurat_clusters"
)
}
\arguments{
\item{so}{seurat object}

\item{refList}{a given reference cell type markers}

\item{clusterMarkerPath}{the path of calculated markers generated by seurat find_markers}

\item{tissue}{The tissue name, used to identify markers in same cell types but from different tissues, such as Epithelial.CRC and Epithlial.LC}

\item{annotateName}{The feature name for given cell type}

\item{clusterName}{The feature name for cluster}
}
\value{
a cluster-marker information table, cluster as subclusters for given cell types with novel markers or cluster contains several given cell types
}
\description{
write a function to test consistency between known annotated labels and seurat clusters
a reference cell markers for known cell type labels
For those cell types with several seurat clusters, to check:
1. Whether those subclusters all express reference markers
2. Is there any significant novel markers for the specific cluster which were overlapped between distinct samples.
Warning clusterName should contains elements the same aligned to those used in the cluster marker file
}
