% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_hull.R
\name{expand_hull}
\alias{expand_hull}
\title{Expand Hull}
\usage{
expand_hull(x, d)
}
\arguments{
\item{x}{a \code{sccm_ch} object}

\item{d}{the amount to move each vertices outward}
}
\description{
Move the vertices of a convex hull outward.
}
