% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{fit2df}
\alias{fit2df}
\title{fit2df}
\usage{
fit2df(fit)
}
\arguments{
\item{fit}{the fitted model generated by causal mediation}
}
\value{
a data.frame with the information of the components used in the mediation relationship and their belonged groups
}
\description{
Change the format of fit model to data.frame
}
