% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_causal_mediation.R
\name{generate_matrix_for_CMA}
\alias{generate_matrix_for_CMA}
\title{generate_matrix_for_CMA}
\usage{
generate_matrix_for_CMA(
  so,
  feature.name,
  int.type,
  dataDir,
  ctlist_to_check,
  funcToUse,
  species = "mouse",
  batch = "orig.ident",
  ifSave = FALSE,
  ifReturn = TRUE,
  outputFile = NULL,
  customedSymbolList = NULL
)
}
\arguments{
\item{so}{a seurat object}

\item{feature.name}{the variable name of cell types used in the seurat object.}

\item{int.type}{the method used for measure outbound interaction strength. 'cellchat.raw' for interaction number estimated by cellchatDB. 'cellchat.p' for cellchat estimated interaction possibility scores. 'iTALK' for iTALK scores.}

\item{dataDir}{the directory of cell chat results}

\item{ctlist_to_check}{a character of cell combinations, each elements contains 3 cell types connected by comma.}

\item{funcToUse}{a list of grouped genes with list names show function names. Obtained from get_functions_to_view.}

\item{species}{the species for the reference database of cellchat database.}

\item{batch}{the name of the batch feature for the seurat object. Default is "orig.ident".}

\item{ifSave}{a boolean variable. save result matrix to outputFile. Default is FALSE.}

\item{ifReturn}{a boolean variable. Only works when ifSave is TRUE. if ifReturn is TRUE, return the maxtrix when save the cma matrix to outputFile. Default is TRUE.}

\item{outputFile}{default is NULL. If ifSave is TRUE and the directory assigned for outputFile doesn't exist, use the current working directory and save the results to mat4cma.rds.}

\item{customedSymbolList}{default is NULL. Pass a named vector for single cell studies with customed defined gene names. The names of the vector are customed defined names, the corresponding values are official gene symbols.}
}
\value{
a data.frame which contains the interaction strength calculated by CellChat
}
\description{
Generate the matrix used for Causal mediation analysis
}
