% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{get_distinct_colorlist}
\alias{get_distinct_colorlist}
\title{get_distinct_colorlist}
\usage{
get_distinct_colorlist()
}
\value{
a list of color list
}
\description{
Get 28 most distinct customed color values
}
