% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{get_endx}
\alias{get_endx}
\title{get_endx}
\usage{
get_endx(x1, y1, x2, y2, height, width, end = "rect")
}
\arguments{
\item{x1}{the x for start position}

\item{y1}{the y for start position}

\item{x2}{the x for end position}

\item{y2}{the y for end position}

\item{height}{the height for the line}

\item{width}{the width for the line}

\item{end}{the target component is rectangle}
}
\value{
the x position for draw
}
\description{
get the end value of x
}
