% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{get_random_colorlist}
\alias{get_random_colorlist}
\title{get_random_colorlist}
\usage{
get_random_colorlist(valList, named = T)
}
\arguments{
\item{valList}{a list of value}

\item{named}{if return the name of input valList}
}
\value{
the generated color list.
}
\description{
Get random generated color list
}
