% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{get_startx}
\alias{get_startx}
\title{get_startx}
\usage{
get_startx(x1, y1, x2, y2, height, width, start = "rect")
}
\arguments{
\item{x1}{the x for start position}

\item{y1}{the y for start position}

\item{x2}{the x for end position}

\item{y2}{the y for end position}

\item{height}{the height for the line}

\item{width}{the width for the line}

\item{start}{the start component is rectangle}
}
\value{
get the x postion for draw
}
\description{
get the start x postion
}
