% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{getdstamp}
\alias{getdstamp}
\title{getdstamp}
\usage{
getdstamp(strlen = 6)
}
\arguments{
\item{strlen}{Default length is 6. Show year, month, date without delimiter.}
}
\value{
a numbered string
}
\description{
get date timestamp
}
