% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexagonalFish.R
\docType{data}
\name{HexagonalFish}
\alias{HexagonalFish}
\title{Hexagonal Fish by Mc Esher}
\format{A data frame
\describe{
  \item{pixel_x}{x coordinate}
  \item{pixel_y}{y coordinate}
}}
\source{
\url{http://www.wikiart.org/en/m-c-escher/not_detected_204748}
}
\usage{
HexagonalFish
}
\description{
A dataset containing (x, y) coordinates (pixels) for
MC Escher's Hexagonal Fish
}
\examples{
data("HexagonalFish", package = "sccm")
summary(HexagonalFish)

}
\keyword{datasets}

