% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_scObj.R
\name{integrateSO}
\alias{integrateSO}
\title{integrateSO}
\usage{
integrateSO(solist)
}
\arguments{
\item{solist}{a list of seurat objects to integrate}
}
\value{
a combined seurat object
}
\description{
integrated seurat object

This is a customed one-step function wrapped the integration pipeline based on functions imported from Seurat and harmony.
}
