% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_anticlockwise.R
\name{is_anticlockwise}
\alias{is_anticlockwise}
\title{Testing for anti-clockwise ordering of points of a convex hull.}
\usage{
is_anticlockwise(x)
}
\arguments{
\item{x}{\code{sccm_pg} object.}
}
\description{
Given a polygon, test if the polygon is a convex hull and if it is, if the
vertices are listed in an anti-clockwise order.
}
\examples{
square1 <- polygon(c(-1, 1, 1, -1), c(-1, -1, 1, 1))
square2 <- polygon(c(1, 1, -1, -1), c(-1, 1, 1, -1))
square3 <- polygon(c(-1, 1, 1, -1), c(1, 1, -1, -1))

is_anticlockwise(square1)
is_anticlockwise(square2)
is_anticlockwise(square3)

}
\seealso{
\code{\link{polygon}} \code{\link{convex_hull}}
\code{\link{is_convex_hull}}
}
