% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(dataDir, dataType, metaPath = NULL)
}
\arguments{
\item{dataDir}{the directory path for single cell data}

\item{dataType}{the format of the raw data. Accepted values are 10x, h5, rds.}

\item{metaPath}{the filepath of the meta file. Default is NULL.}
}
\value{
a seurat object
}
\description{
import the single cell data and transfer data to seurat object.
}
