% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{load_file}
\alias{load_file}
\title{load_file}
\usage{
load_file(filepath, delimiter = NULL, ...)
}
\arguments{
\item{filepath}{file path of the data table}

\item{delimiter}{Default is NULL.}
}
\value{
data table
}
\description{
load table files
}
