% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{load_packages}
\alias{load_packages}
\title{load_packages}
\usage{
load_packages(packs_to_check)
}
\arguments{
\item{packs_to_check}{a vector of packages need to be checked}
}
\description{
Loading the packages required by scCellMediator
}
