% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{load_so}
\alias{load_so}
\title{load_so}
\usage{
load_so(so_dir)
}
\arguments{
\item{so_dir}{The directory of stored seurat object in rds format}
}
\value{
seurat object
}
\description{
load newest seurat object
}
