% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellchat.R
\name{loadCellChatDB}
\alias{loadCellChatDB}
\title{loadCellChatDB}
\usage{
loadCellChatDB(species)
}
\arguments{
\item{species}{current legal species are human, mouse, zebrafish.}
}
\value{
a cell database or NULL
}
\description{
load CellChat reference database for a specific species
}
