% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{mergeDataPos}
\alias{mergeDataPos}
\title{mergeDataPos}
\usage{
mergeDataPos(res, df2, whatLabels = "est", width = 5, height = 3)
}
\arguments{
\item{res}{the orignal result}

\item{df2}{the filtered matrix}

\item{whatLabels}{default est for estimation}

\item{width}{default 5}

\item{height}{default 3}
}
\description{
merge the postion of the summary table
}
