% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{named_vec}
\alias{named_vec}
\title{named_vec}
\usage{
named_vec(inames, ivalues)
}
\arguments{
\item{inames}{A vector of name list.}

\item{ivalues}{A vector of value list.}
}
\value{
A named vector.
}
\description{
Generating a named vector.
}
