% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_causal_mediation.R
\name{normalized_interaction_strength}
\alias{normalized_interaction_strength}
\title{normalized_interaction_strength}
\usage{
normalized_interaction_strength(int.scores)
}
\arguments{
\item{int.scores}{the vector for interaction scores.}
}
\value{
the normalized interaction score. New range: 0-1.
}
\description{
normalized_interaction_strength
}
