% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2d.R
\name{p2d}
\alias{p2d}
\title{Polygon to Disk}
\usage{
p2d(.data, pg = sccm::convex_hull(.data), ...)
}
\arguments{
\item{.data}{the data set to map to the disk}

\item{pg}{the polygon containing the data, the default is the convex
hull about \code{.data}.}

\item{...}{additional arguments passed to \code{scmap}}
}
\value{
a \code{sccm_p2d} object, a list with the following elements:
\describe{
\item{mapped}{a n by 2 matrix with the (x, y) coordinates of the mapped data.}
\item{polygon}{The polygon the .data was contained in}
\item{data}{The original data}
\item{mapping}{the Schwarz-Christoffel mapping used.}
}
}
\description{
Map the data within a polygon to the unit disk via Schwarz-Christoffel
conformal mapping.
}
\seealso{
\code{\link{scmap}} \code{\link{d2p}} \code{\link{p2p}}
}

