% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2p.R
\name{p2p}
\alias{p2p}
\title{Polygon to Polygon}
\usage{
p2p(
  .data,
  pg1 = sccm::convex_hull(.data),
  pg2 = sccm::polygon(x = c(-1, 1, 1, -1), y = c(-1, -1, 1, 1)),
  ...
)
}
\arguments{
\item{.data}{the data set to map}

\item{pg1}{the polygon containing the \code{.data}}

\item{pg2}{the polygon to map \code{.data} to.}

\item{...}{additional arguments passed to \code{scmap}}
}
\value{
a \code{sccm_p2d} object, a list with the following elements:
\describe{
\item{mapped}{a n by 2 matrix with the (x, y) coordinates of the mapped data
in the target polygon.}
\item{disked}{a n by 2 matrix with the (x, y) coordinates of the mapped data
in the unit disk.}
\item{data}{The original data}
\item{pg1}{The polygon the .data was contained in}
\item{pg2}{The polygon the .data was mapped into}
\item{p2d}{the Schwarz-Christoffel mapping between the unit disk and polygon1.}
\item{d2p}{the Schwarz-Christoffel mapping between the unit disk and polygon2.}
}
}
\description{
Map the data within a polygon to another polygon  via Schwarz-Christoffel
conformal mapping.
}
\examples{
set.seed(42)
n <- 2500L
dat <- data.frame(x = 1 + runif(n)) 
dat$y <- log(dat$x) + rnorm(n, sd = 0.2)
ch <- sccm::convex_hull(dat)
plot(ch)

trans <- sccm::p2p(dat)
plot(trans)

star <- 
   rbind(sccm::polar2cartesian(r = 1.0, theta = seq(0, 1.6, by = 0.4) * pi),
         sccm::polar2cartesian(r = 0.6, theta = seq(0.2, 1.8, by = 0.4) * pi))
star <- star[rep(1:5, each = 2) + rep(c(0, 5), times = 5), ]

trans <- sccm::p2p(dat, pg2 = sccm::polygon(star))
plot(trans)


}
\seealso{
\code{\link{scmap}} \code{\link{d2p}} \code{\link{p2d}}
}
