% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{paste_with_sep}
\alias{paste_with_sep}
\title{paste_with_sep}
\usage{
paste_with_sep(eles, delimiter = ",")
}
\arguments{
\item{eles}{Elements to link together}

\item{delimiter}{The delimiter to use (default is ",").}
}
\value{
A string with elements linked by selected delimiter.
}
\description{
Gluing elements together with the given delimiter.
}
