% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polar2cartesian}
\alias{polar2cartesian}
\alias{cartesian2polar}
\title{Translate between Polar and Cartesian Coordinates.}
\usage{
polar2cartesian(r, theta)

cartesian2polar(x, y)
}
\arguments{
\item{r}{a numeric vector of radii}

\item{theta}{a numeric vector of angles (in radians)}

\item{x}{a numeric vector of x coordinates}

\item{y}{a numeric vector of y coordinates}
}
\value{
A numeric matrix
}
\description{
Given (x,y) return (r, theta), or visa-versa.
}
