% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{polygon}
\alias{polygon}
\title{Polygon}
\usage{
polygon(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates, a n by 2 matrix, or anything that
can be coerced to a matrix via \code{as.matrix}.}

\item{y}{a numeric vector of y coordinates, only needed if x is a numeric
vector}
}
\value{
A list
}
\description{
return the polygon and external angles for a set of (x, y) points
}

