% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate}
\usage{
rotate(x, counter = TRUE, times = 1L)
}
\arguments{
\item{x}{a \code{sccm_ch} or \code{sccm_pg} object}

\item{counter}{(logical, default = \code{TRUE}), rotate counter clockwise}

\item{times}{(integer, default = 1), number vertices to rotate through}
}
\description{
Rotate a polygon or convex hull
}
