% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_causal_mediation.R
\name{run_CausalMediation}
\alias{run_CausalMediation}
\title{run_CausalMeidation}
\usage{
run_CausalMediation(
  mat4mca,
  useColor = TRUE,
  colList = NULL,
  pvalue = 0.001,
  ifPlot = TRUE,
  ifSavePlot = TRUE,
  ifCombinePlot = TRUE,
  ifReturnPlots = FALSE,
  ifReturnFitModels = FALSE,
  ifSaveFitModels = TRUE,
  ifPlotOnlySameDirect = TRUE,
  datalabel = "",
  saveDir = NULL
)
}
\arguments{
\item{mat4mca}{a matrix generated by generate_matrix_for_CMA using the seurat object}

\item{useColor}{a boolean value.}

\item{colList}{a vector of color values. Default is NULL.}

\item{pvalue}{the threshold value to determine the significance.}

\item{ifPlot}{a boolean value. Whether to generate the plots for those cells displayed a partial/fully mediation relationship. We only generate plots for those show a consistent direction of mediation effects between Sender-Mediator pair and Mediator-Receiver pair. Default is TRUE.}

\item{ifSavePlot}{a boolean value. Masked by ifPlot. Only works when ifPlot is True.}

\item{ifCombinePlot}{a boolean value. Masked by ifSavePlot. Only works when ifSavePlot is True. If many causual mediation relationships detected, we recommend to display the relationships in a combined way. Each combined figure can display up to nine relationships.}

\item{ifReturnPlots}{a boolean value. Masked by ifPlot. Only works when ifPlot is True. Whether to print the plots in R IDE. If not, save them to saveDir with the prefix, . Otherwise,}

\item{ifReturnFitModels}{a boolean value.}

\item{ifSaveFitModels}{a boolean value.}

\item{ifPlotOnlySameDirect}{a boolean value. Default is TRUE. If true, only draw relationships that have same regulation directions in 'Sender->Mediator' and 'Mediator->Receiver'.}

\item{datalabel}{a character value. Default is "". To label saved fitted model with a customed name.}

\item{saveDir}{the path for the directory to save plots and fitted models.}
}
\value{
return NULL or a list of results with fitted models and/or plots of accepted causal mediation relationships.
}
\description{
run causual mediation analysis

causal mediation based on specific ligand-receptor interactions
deparacated parameter: mod the mode used to measure the signals. spc for specific signals, which only consider the possible interactions between the interested cells. gnl for general signals, which consider all signals of ligands and receptors from the cells.
}
