% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{run_combined_seurat}
\alias{run_combined_seurat}
\title{run_combined_seurat}
\usage{
run_combined_seurat(
  solist_dir,
  save_dir = NULL,
  ifReturn = FALSE,
  filterRules = NULL,
  acceptRules = NULL,
  dimnum = 50,
  k.anchor = 5,
  k.filter = 20,
  k.weight = 10
)
}
\arguments{
\item{solist_dir}{the path of directory to the directories of stored seurat objects}

\item{save_dir}{the output directory of integrated object.}

\item{filterRules}{default is NULL. Use the rules to filter seurat objects before the integration. Using a string like 'CellType %in% c("EC","T_cell","Cancer") & stringr::str_detect(.$CellTypeStatus, "CT_confirmed") & nCount_RNA > 500 & nFeature_RNA > 200' to filter the result.}

\item{acceptRules}{default is NULL. Use the rules to determine whether to include a filtered seurat object or not. Syntax only work with 'CellType,>,50' to determine.}

\item{dimnum}{the maximum dimension to use}
}
\description{
run combined seurat object
}
