% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_iTALK.R
\name{run_iTALK}
\alias{run_iTALK}
\title{run_iTALK}
\usage{
run_iTALK(
  so,
  feature.name = "ct.major",
  batch.name = "orig.ident",
  output = NULL,
  useStoredResampledSO = F,
  storedResampledSODir = NULL,
  minCNUM.tot = 10,
  sampleProp = 0.5,
  minCNUM.new = 10,
  runs = 10,
  storeResampledSO = FALSE,
  customedSymbolList = NULL
)
}
\arguments{
\item{so}{a seurat object}

\item{feature.name}{the feature used to control sampling cell proportion}

\item{batch.name}{the feature used to separate cells from different resources}

\item{output}{the directory path to store results}

\item{useStoredResampledSO}{a boolean value. Default is FALSE. Set to TRUE to skip sample resampling by offering stored resampled seurat objects.}

\item{storedResampledSODir}{the path to stored resampled seurat objects. When useStoredResampledSO is TRUE, load the stored resampled seurat objects.}

\item{minCNUM.tot}{default is 10. The minimum number of cells in each cell cluster per feature per batch. Unsatisfied clusters will be removed.}

\item{sampleProp}{default is 0.5. Value ranges between 0-1. To Estimate the proportion of resampled cell number in each cell cluster.}

\item{minCNUM.new}{default is 10. The minimum number of cells in each resampled cell clusters per feature per batch. Unsatisfied clusters will be removed.}

\item{runs}{default is 10. The replicate number of resampled times.}

\item{storeResampledSO}{default is FALSE. a boolean value. If TRUE, store resampled seurat objects to output.}

\item{cutomedSymbolList}{default is NULL. Pass a named vector for single cell studies with customed defined gene names. The names of the vector are customed defined names, the corresponding values are official gene symbols.}
}
\description{
run gene expression to gain LR interactions
}
