% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{run_seurat}
\alias{run_seurat}
\title{run_seurat}
\usage{
run_seurat(
  so,
  label = "",
  batch = NULL,
  features_to_view = NULL,
  save = T,
  so_dir = NULL,
  prior_metadata = NULL,
  addNamedColorList = NULL,
  treat.cellname = F,
  mtlabel = "MT-",
  min_avg = 0.01,
  assay = "RNA",
  scale_factor = 1e+06,
  min_bin = 50,
  min_cutoff = c(0.01, Inf),
  calcClustMarker = T,
  dimnum = "auto",
  dim_sd = 1.25,
  dim_pval = 0.01,
  clustRes = 0.8
)
}
\arguments{
\item{so}{a seurat object.}

\item{label}{the annotation of seurat obejct.}

\item{batch}{the variable name of batch.}

\item{features_to_view}{default is NULL. a list of features for the visualization of feature distribution in the UMAP.}

\item{save}{a boolean value. Whether to save treated seurat object to so_dir. Default is FALSE.}

\item{so_dir}{default is NULL. The directory to store treated seurat object.}

\item{prior_metadata}{default is NULL. a customed meta table.}

\item{treat.cellname}{a boolean value. If TRUE, add label to seurat cell names.}

\item{mtlabel}{default is "MT-". You can define a customed mitochondrial gene list by yourself.}

\item{min_avg}{default is 0.01. Minimum expressed average gene.}

\item{assay}{default is "RNA".}

\item{scale_factor}{default is 1e+06 for CPM.}

\item{min_bin}{default is 50. Minimum bin value.}

\item{min_cutoff}{default is c(0.01, Inf).}

\item{dimnum}{default is "auto".}
}
\description{
a customed pipeline for the preprocessing of seurat object.
}
