% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_settings.R
\name{savePlot}
\alias{savePlot}
\title{savePlot}
\usage{
savePlot(
  pObj,
  pDir = ".",
  pType = "ggplot",
  pName = "",
  pWd = 7,
  pHt = 7,
  pFmt = "pdf",
  dStamp = T
)
}
\arguments{
\item{pObj}{the ggplot2 object}

\item{pDir}{the path of directory to output.}

\item{pType}{types of plot, such as barplot, ggplot, circleplot, etc.}

\item{pName}{a recognizable name to the plot}

\item{pWd}{the width of the output plot. Default is 7.}

\item{pHt}{the height of the output plot. Default is 7.}

\item{pFmt}{the format of the figure. Default is .pdf.}

\item{dStamp}{a boolean value to decide whether add time stamp to the output filename.}
}
\description{
save ggplot2 object
}
