% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{seekGroup1}
\alias{seekGroup1}
\title{seekGroup1}
\usage{
seekGroup1(var, res)
}
\arguments{
\item{var}{a character of a component name in result table}

\item{res}{a data frame contains the treated summary table of the mediation model}
}
\value{
the belonged group of the varaible: "X", "Y", "M1", "M2"
}
\description{
condition1 for seek group
}
