% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{seekGroup2}
\alias{seekGroup2}
\title{seekGroup2}
\usage{
seekGroup2(var, res, group)
}
\arguments{
\item{var}{a character of a component name in result table}

\item{res}{a data frame contains the treated summary table of the mediation model}

\item{group}{a vector contains names of the components used in the mediation model}
}
\value{
the position for component "H"
}
\description{
add position for Hidden component
}
