% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{separate_so}
\alias{separate_so}
\title{separate_so}
\usage{
separate_so(so, batch = "orig.ident")
}
\value{
a list of seurat subobjects
}
\description{
separate seurat object by batch feature
}
