% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relationships.R
\name{theme_clean}
\alias{theme_clean}
\title{output a cleaned theme for ggplot object}
\usage{
theme_clean(base_size = 12, base_family = "Arial")
}
\arguments{
\item{base_size}{the font size used, default 12}

\item{base_family}{the font family used, default Arial}
}
\value{
the set theme
}
\description{
remove the title, ticks, and text on both axises
remove background grid lines and color rectangles
compress the margin of the figure
}
