% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_so_preprocessing.R
\name{update_so}
\alias{update_so}
\title{update_so}
\usage{
update_so(so, so_dir, meta.only = T)
}
\arguments{
\item{so}{seurat object}

\item{so_dir}{the seurat directory}

\item{meta.only}{a boolean variable. Default is TRUE. If TRUE, only update metatable, seurat object will not be updated.}
}
\description{
save or update seurat object and its meta.data
}
