% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_grid.R
\name{conformal_grid}
\alias{conformal_grid}
\title{Conformal Grid}
\usage{
conformal_grid(x, ubreaks = 7, vbreaks = 7, n = 100)
}
\arguments{
\item{x}{a \code{sccm_pg} or \code{sccm_ch} object}

\item{ubreaks, vbreaks}{Either an integer value or numeric vector for the
location of the grid lines.  If an integer value is given,
then the dimension will have equidistant grid lines.  Else, specify specific
values in the u and/or v space between -1, and 1.}

\item{n}{number of data points to use on each margin of the grid.}
}
\description{
Given a polygon, find the orthogonal grid from a square projected into the
polygon.
}
\details{
Given a orthogonal grid in (u, v) space, \code{conformal_grid} finds and
allows one to plot the mapped grid into the polygon in (x, y) space.  The (u,
v) space is a square with vertices at (-1, -1), (1, -1), (1, 1), and (-1, 1).
}
