% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Convex Hull}
\usage{
convex_hull(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates, a n by 2 matrix, or anything that
can be coerced to a matrix via \code{as.matrix}.}

\item{y}{a numeric vector of y coordinates, only needed if x is a numeric
vector}
}
\value{
a list
}
\description{
return the convex hull for a set of (x, y) points
}

