% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{convex_hull}
\alias{convex_hull}
\title{Convex Hull}
\usage{
convex_hull(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates, a n by 2 matrix, or anything that
can be coerced to a matrix via \code{as.matrix}.}

\item{y}{a numeric vector of y coordinates, only needed if x is a numeric
vector}
}
\value{
A \code{sccm_ch} object.  This is a \code{list} with the following
elements:
\describe{
\item{vertices}{a n by 2 matrix of the vertices defining the convex
hull.}
\item{data}{a matrix of the data provided}
\item{beta}{a numeric vector of the exterior angles for each of the
vertices.}
\item{indices}{a numeric vector of the row indices from \code{x}
corresponding to the data values found to be the vertices of the convex
hull.}
}
}
\description{
Return the convex hull for a set of (x, y) points
}
\details{
The \code{convex_hull} function returns the vertices, in an anti-clockwise
order, of the convex hull encompassing the provided data.  This differs from
the \code{polygon} function as convex hull is generated automatically via
Andrew's Monotone algorithm and the \code{polygon} function only provides the
needed angles between edges of a given set of points.
}
\examples{

set.seed(42)
dat <- data.frame(x = rnorm(100), y = rexp(100))
ch <- convex_hull(dat)
ch
plot(ch) 

}
\seealso{
\code{\link{polygon}} \code{\link{p2d}} \code{\link{d2p}}
\code{\link{p2p}}
}

