% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2d.R
\name{p2d}
\alias{p2d}
\title{Polygon to Disk}
\usage{
p2d(.data, pg = sccm::convex_hull(.data), ...)
}
\arguments{
\item{.data}{the data set to map to the disk}

\item{pg}{the polygon containing the data, the default is the convex
hull about \code{.data}.}

\item{...}{additional arguments passed to \code{scmap}}
}
\value{
a list
}
\description{
Map the data within a polygon to the unit disk via Schwarz-Christoffel
conformal mapping.
}
\seealso{
\code{scmap}
}

